# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "The Reward Loyalty API provides programmatic access to the Reward Loyalty platform, enabling integration with external systems, mobile apps, and custom workflows. This RESTful API supports authentication, loyalty card management, member operations, partner operations, and transaction processing.\n\n## Authentication\n\nAll authenticated endpoints use Bearer token authentication via Laravel Sanctum. Include your token in the Authorization header.\n\nTokens are obtained by calling the login endpoints for each user type (Admin, Partner, or Member).\n\n## Rate Limiting\n\nAPI requests are rate-limited to ensure fair usage. Standard limits apply per authenticated user.\n\n## Response Format\n\nAll responses are returned in JSON format. Successful responses include the requested data, while errors include a message field describing the issue.",
        "contact": {
            "name": "Reward Loyalty Website",
            "url": "https://distech.co.za"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/member": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "Retrieve member data",
                "description": "Fetches the data of the authenticated member.",
                "operationId": "getMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
...
        "/{locale}/v1/partner/members": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "List all members",
                "description": "Retrieve all members registered in the partner's clubs.",
                "operationId": "getPartnerMembers",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Members retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Member"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Partner"
                ],
                "summary": "Create a new member",
                "description": "Register a new member in one of the partner's clubs.",
                "operationId": "createPartnerMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "club_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Member name",
                                        "type": "string",
                                        "example": "Jane Customer"
                                    },
                                    "email": {
                                        "description": "Member email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "password": {
                                        "description": "Password (optional, auto-generated if not provided)",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "club_id": {
                                        "description": "Club ID",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019b8d49-1234-7abc-8def-17f2107a8ec0"
                                    },
                                    "accepts_emails": {
                                        "description": "Accepts marketing emails",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Member created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/partner/members/{memberId}": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Get member details",
                "description": "Retrieve details of a specific member by ID.",
                "operationId": "getPartnerMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Member not found"
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Partner"
                ],
                "summary": "Update a member",
                "description": "Update an existing member's details.",
                "operationId": "updatePartnerMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated@example.com"
                                    },
                                    "password": {
                                        "description": "New password (optional)",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "accepts_emails": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Member updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Member not found"
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Partner"
                ],
                "summary": "Delete a member",
                "description": "Delete a member. This action cannot be undone.",
                "operationId": "deletePartnerMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "memberId",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Member deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Member not found"
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
...
        "/{locale}/v1/staff/member/{identifier}": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "Find a member",
                "description": "Look up a member by ID, unique identifier, email, or phone. Returns member details and balances.",
                "operationId": "staffFindMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Member ID, unique identifier, email, or phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Member not found"
                    }
                },
                "security": [
                    {
                        "staff_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Response when authentication is required but not provided or invalid",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Authentication error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "Member": {
                "description": "Loyalty program member (customer) account",
                "required": [
                    "id",
                    "unique_identifier",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier",
                        "type": "string",
                        "format": "uuid",
                        "example": "019b8d48-97df-7177-8aa0-47f2107a8eb3"
                    },
                    "unique_identifier": {
                        "description": "Human-readable unique ID",
                        "type": "string",
                        "example": "700-857-223-945"
                    },
                    "name": {
                        "description": "Member name",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Jane Customer"
                    },
                    "email": {
                        "description": "Email address",
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "email_verified_at": {
                        "description": "Email verification timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "locale": {
                        "description": "Preferred locale",
                        "type": "string",
                        "example": "en_US",
                        "nullable": true
                    },
                    "currency": {
                        "description": "Preferred currency",
                        "type": "string",
                        "example": "USD",
                        "nullable": true
                    },
                    "time_zone": {
                        "description": "Time zone",
                        "type": "string",
                        "example": "America/New_York",
                        "nullable": true
                    },
                    "accepts_emails": {
                        "description": "Whether member accepts marketing emails",
                        "type": "boolean",
                        "example": true
                    },
                    "number_of_times_logged_in": {
                        "description": "Total login count",
                        "type": "integer",
                        "example": 12
                    },
                    "last_login_at": {
                        "description": "Last login timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Account creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last modification timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "avatar": {
                        "description": "URL to avatar image",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "admin_auth_token": {
                "type": "http",
                "description": "Administrator authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "partner_auth_token": {
                "type": "http",
                "description": "Partner authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Member authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}